+++ asusd/src/ctrl_anime/mod.rs
@@ -24,7 +24,11 @@ impl GetSupported for CtrlAnime {
     type A = AnimeSupportedFunctions;
 
     fn get_supported() -> Self::A {
-        AnimeSupportedFunctions(HidRaw::new("193b").is_ok())
+        if USBRaw::new(0x193b).is_ok() {
+            AnimeSupportedFunctions(true)
+        } else {
+            AnimeSupportedFunctions(HidRaw::new("193b").is_ok())
+        }
     }
 }
 
@@ -63,8 +67,8 @@ pub struct CtrlAnime {
 impl CtrlAnime {
     #[inline]
     pub fn new(config: AnimeConfig) -> Result<CtrlAnime, RogError> {
-        let hid = HidRaw::new("193b").ok();
         let usb = USBRaw::new(0x193b).ok();
+        let hid = HidRaw::new("193b").ok();
         let node = if usb.is_some() {
             unsafe { Node::Usb(usb.unwrap_unchecked()) }
         } else if hid.is_some() {