+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixes to asusd to ensure fan curve defaults are loaded if the config file fails
 - Further refine the asusctl CLI for fan-curve control
 - Fixes to AniMe detection
+- Fixes to aura config creation/loading
 
 ### Added
 - Support for GV601V LED modes