+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [v4.7.1]
 ### Changed
 - Fixes to asusctl CLI tool to show fan curves
 - Fixes to asusd to ensure fan curve defaults are loaded if the config file fails